Add simple honey-pot based spam protection to contact form
The spam protection is meant to filter out very simple spam bots by
assuming they will fill out a (hidden) website-field.

If a user accidentally fills out the field---e.g., through a browser
extension---the UI will return a message to tell them to not fill
out the website-field. Bots smart enough to catch this are anyway
smart enough not to fall for a honey pot.